src/test/javascript/spec/app/entities/professional-profile-user/professional-profile-user.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import * as moment from 'moment';
import { DATE_TIME_FORMAT } from 'app/shared/constants/input.constants';
import { ProfessionalProfileUserService } from 'app/entities/professional-profile-user/professional-profile-user.service';
import { IProfessionalProfileUser, ProfessionalProfileUser } from 'app/shared/model/professional-profile-user.model';

describe('Service Tests', () => {
  describe('ProfessionalProfileUser Service', () => {
    let injector: TestBed;
    let service: ProfessionalProfileUserService;
    let httpMock: HttpTestingController;
    let elemDefault: IProfessionalProfileUser;
    let expectedResult: IProfessionalProfileUser | IProfessionalProfileUser[] | boolean | null;
    let currentDate: moment.Moment;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(ProfessionalProfileUserService);
      httpMock = injector.get(HttpTestingController);
      currentDate = moment();

      elemDefault = new ProfessionalProfileUser(0, 0, 0, 'AAAAAAA', currentDate, false);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign(
          {
            creationDate: currentDate.format(DATE_TIME_FORMAT),
          },
          elemDefault
        );

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a ProfessionalProfileUser', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
            creationDate: currentDate.format(DATE_TIME_FORMAT),
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            creationDate: currentDate,
          },
          returnedFromService
        );

        service.create(new ProfessionalProfileUser()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a ProfessionalProfileUser', () => {
        const returnedFromService = Object.assign(
          {
            profesionalType: 1,
            pricePerHour: 1,
            description: 'BBBBBB',
            creationDate: currentDate.format(DATE_TIME_FORMAT),
            state: true,
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            creationDate: currentDate,
          },
          returnedFromService
        );

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of ProfessionalProfileUser', () => {
        const returnedFromService = Object.assign(
          {
            profesionalType: 1,
            pricePerHour: 1,
            description: 'BBBBBB',
            creationDate: currentDate.format(DATE_TIME_FORMAT),
            state: true,
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            creationDate: currentDate,
          },
          returnedFromService
        );

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a ProfessionalProfileUser', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
